% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaning-functions.R
\name{char_to_num}
\alias{char_to_num}
\title{Character to Numeric Data Frame}
\usage{
char_to_num(df)
}
\arguments{
\item{df}{The vius data frame with the character columns to be transformed.}
}
\value{
The vius data with numeric vectors
}
\description{
Takes certain character columns in the vius data and transforms them into
numeric.
}
