% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaning-functions.R
\name{convert_names}
\alias{convert_names}
\title{Convert Codes to correct charater values}
\usage{
convert_names(df)
}
\arguments{
\item{df}{The vius data frame with codes for variables, must not include NAs}
}
\value{
The vius data with useful columns
}
\description{
Converts the VIUS specific codes for certain categorical variables into user
understandable characters.
}
