% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaning-functions.R
\name{drop_bad_values}
\alias{drop_bad_values}
\title{Drop Bad Values from a Data Frame}
\usage{
drop_bad_values(df)
}
\arguments{
\item{df}{A vius dataframe to be cleaned}
}
\value{
The cleaned vius dataframe
}
\description{
This function scans the provided data frame for values in vius data sets that
are not useful and removes them.
}
