% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaning-functions.R
\name{drop_cols}
\alias{drop_cols}
\title{Drop columns not used in the VIUSr package}
\usage{
drop_cols(df)
}
\arguments{
\item{df}{The vius data frame}
}
\value{
The vius data frame with only the desired columns
}
\description{
This function takes the raw vius data and drops all columns in the data set
that are not used by the VIUSr package.
}
