% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodyType_vius_plotter.R
\name{get_btype_graphs}
\alias{get_btype_graphs}
\title{Plot Weighted Averages by Body Type}
\usage{
get_btype_graphs(
  vius1,
  var,
  plot_title = "BAR CHART",
  x_plot_label = "Variable Average",
  y_plot_label = "Variable of Interest"
)
}
\arguments{
\item{vius1}{A data frame containing at least the columns \code{TABWEIGHT},
\code{BTYPE}, and the variable of interest.}

\item{var}{The column of the VIUS dataset. Must be a column of numeric
values in the cleaned VIUS dataset available within the VIUSr package.
Options are AVGWEIGHT (average vehicle weight), ER_COST (average cost of
extensive vehicle repairs), GM_COST (average cost of general vehicle
maintenance), MILESANNL (average miles traveled per vehicle in the year
of the survey), MILESLIFE (average miles traveled per vehicle since the
vehicle was manufactured), MPG (average miles per gallon), TRIPOFFROAD
(The average percentage of miles driven off-road per vehicle)}

\item{plot_title}{A character string for the plot title.
Defaults to \code{"Title"}.}

\item{x_plot_label}{A character string for the x-axis label.
Defaults to \code{"x axis"}.}

\item{y_plot_label}{A character string for the y-axis label.
Defaults to \code{"y axis"}.}
}
\value{
A ggplot2 object representing the bar chart.
}
\description{
This function calculates the weighted average of a given variable across
vehicle body types using \code{TABWEIGHT} for weights, and generates a
horizontal bar chart. Each bar represents a body type, colored
individually, and sorted by the weighted average.
}
\examples{
\dontrun{
get_btype_graphs(vius1, MPG,
  plot_title = "Average MPG by Body Type",
  x_plot_label = "Body Type", y_plot_label = "Average MPG")
}

}
