% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statewise_vius_plotter.R
\name{get_state_graphs}
\alias{get_state_graphs}
\title{Generate State-Level Bar Graphs from VIUS Data}
\usage{
get_state_graphs(
  dataset,
  db_header,
  states = NULL,
  plot_title = "State - ",
  x_plot_label = "x axis",
  y_plot_label = "y axis",
  sleep_time = 2
)
}
\arguments{
\item{dataset}{The cleaned VIUS object to be mapped}

\item{db_header}{A column from the dataset used for grouping.}

\item{states}{A vector of state abbreviations or FIPS codes to include in
the plots. If \code{NULL} (default), plots will be generated for all available
states in the dataset.}

\item{plot_title}{A character string used as the title prefix for each
state-level plot.}

\item{x_plot_label}{A character string specifying the label for the x-axis
(which will appear vertically due to \code{coord_flip}).}

\item{y_plot_label}{A character string specifying the label for the y-axis.}

\item{sleep_time}{Time in seconds to pause between plots. Default is 2
seconds.}
}
\value{
No return value. The function produces plots
}
\description{
This function reads the VIUS survey files, filters and summarizes the data
by state and a user-specified column, then creates horizontal bar plots of
estimated vehicle counts for each state.
}
