% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statewise_vius_plotter.R
\name{get_state_maps}
\alias{get_state_maps}
\title{Create a heatmap of the Unites States}
\usage{
get_state_maps(vius, var, var_label = "Value", dollars = FALSE)
}
\arguments{
\item{vius}{The VIUS dataset. This should be a reference to the cleaned VIUS
dataset available within the VIUSr package.}

\item{var}{The column of the VIUS dataset. Must be a column of numeric
values in the cleaned VIUS dataset available within the VIUSr package.
Options are AVGWEIGHT (average vehicle weight), ER_COST (average cost of
extensive vehicle repairs), GM_COST (average cost of general vehicle
maintenance), MILESANNL (average miles traveled per vehicle in the year
of the survey), MILESLIFE (average miles traveled per vehicle since the
vehicle was manufactured), MPG (average miles per gallon), TRIPOFFROAD
(The average percentage of miles driven off-road per vehicle)}

\item{var_label}{A string representing the label the user would like to see
on the map for the value of the selected variable. If no value is entered,
the label will simply say "value."}

\item{dollars}{A boolean value. Defaults to False. If true, will append a "$"
character to the value when displayed on the map.}
}
\value{
A plotly map of the United States with darker states representing
higher values of the selected variable.
}
\description{
This function creates an interactive map of the United States. The color of
each state in the map will depend on a variable available in the VIUS
dataset. The user can select which variable to map out across the United
States and the label for the variable. Then, when the user hovers over a
state with their cursor, the map will display the state and the
corresponding value of the variable for that state.
}
\examples{
get_state_maps(vius, var = MILESANNL, var_label = "Average annual miles")
get_state_maps(vius,
  var = GM_COST,
  var_label = "Average amount spent on general maintenance",
  dollars = TRUE
)
}
