% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statewise_vius_plotter.R
\name{getStateGraphs}
\alias{getStateGraphs}
\title{Generate State-Level Bar Graphs from VIUS Data}
\usage{
getStateGraphs(
  dataset,
  dbHeader,
  states = NULL,
  plotTitle = "State - ",
  xPlotLabel = "x axis",
  yPlotLabel = "y axis",
  sleepTime = 2
)
}
\arguments{
\item{dbHeader}{A column from the dataset used for grouping.}

\item{states}{A vector of state abbreviations or FIPS codes to include in
the plots. If \code{NULL} (default), plots will be generated for all available
states in the dataset.}

\item{plotTitle}{A character string used as the title prefix for each
state-level plot.}

\item{xPlotLabel}{A character string specifying the label for the x-axis
(which will appear vertically due to \code{coord_flip}).}

\item{yPlotLabel}{A character string specifying the label for the y-axis.}

\item{sleepTime}{Time in seconds to pause between plots. Default is 2
seconds.}
}
\value{
No return value. The function produces plots
}
\description{
This function reads the VIUS survey files, filters and summarizes the data
by state and a user-specified column, then creates horizontal bar plots of
estimated vehicle counts for each state.
}
