% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaning-functions.R
\name{select_variables_vius}
\alias{select_variables_vius}
\title{Selects only the wanted variables}
\usage{
select_variables_vius(vius)
}
\arguments{
\item{vius}{The vius data frame all the variables in the public use file}
}
\value{
The vius data with only variables of interest.
}
\description{
In order to get a datset that is much easier, especially for less powerful
machines, to work with
}
