% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vius_raw_sample}
\alias{vius_raw_sample}
\title{The first 100 rows of the VIUS set data with no cleaning done}
\format{
\subsection{\code{vius_raw_sample}}{

A data frame with 100 rows and 168 columns
}
}
\source{
<https://www.census.gov/data/datasets/2021/econ/vius/
2021-vius-puf.html>
}
\usage{
vius_raw_sample
}
\description{
This is a subset of the VIUS data set with none of the package's data
cleaning functions applied to it. Only the first 100 rows of the data were
included to keep the size of the data frame small.
}
\keyword{datasets}
